PHP 8.1: fix tests for trait method replacement
Calling static trait members directly instead via a class using the trait is deprecated as of PHP 8.1.
The functionality will still work until PHP 9.0.

With that in mind, I'm only adjusting the tests to silence the deprecation notice from the direct calls to trait methods.
I've chosen to do so by moving the problematic calls to a separate statement and silencing the method call there. This prevent an failing assertion accidentally being silenced and the test not reporting correctly.

Fixes 2 tests/10 warnings along the lines of:
```
Deprecated: Calling static trait method FooTrait::speak is deprecated, it should only be called on a class using the trait
```

Refs:
* https://wiki.php.net/rfc/deprecations_php_8_1#accessing_static_members_on_traits