Avoid parsing ini files from `parse_ini_file()`
In PHP 7 and 8, `parse_ini_file()` calls the wrapper and sets
STREAM_OPEN_FOR_INCLUDE. None the less, we don't want to parse it.

It seems there's no direct way to differentiate such a call from a
`require` or `include` that we do need to process, other than by
checking the stack.

Fixes antecedent#124