GH Actions: start running tests against PHP 8.2
PHP 8.2 - yet again - contains quite a number of deprecations which will become fatals in PHP 9.0.

As the impact is expected to be large, it is vital that packages used in the CI/test chain are ready in time to allow the applications/frameworks etc using these packages to get ready as well.

With that in mind, I'm proposing to start running the tests against PHP 8.2. At this time, the tests are passing, but keeping an eye on them over the next few months is strongly recommended.