GH Actions: fix test runs on PHP < 7.2
The `setup-php` action runner will automatically install the appropriate version of Composer depending on the PHP version requested - Composer 2.2 for PHP < 7.2 and Composer 2.3+ for PHP 7.2+.

However, using `tools: none` disables this and falls back to the Composer version which is shipped with the Ubuntu image being used.

Composer 2.3, which was released end of March, has a minimum PHP version of PHP 7.2, so will not work/be available to PHP < 7.2, so we need to ensure that the `setup-php` action runner installs the correct version, otherwise the `composer install` needed for the tests will fail.

Removing the `tools: none` should fix this.

**Note:**
Autoload files generated with Composer 2.3 are no longer compatible with PHP < 5.6, though Composer 2.2 (with a minimum PHP version of PHP 5.3) is now a LTS (long term service) release and will receive essential bug fixes.

As the humbug/box config is explicitly set up to not include a Composer generated autoload file, this should not impact the PHAR file which is generated on releases of this project.

If this would ever change, the release workflow would need to be adjusted to include `tools: composer:2.2` for the `setup-php` action runner.

Refs:
* https://blog.packagist.com/composer-2-3/
* https://github.com/composer/composer/releases
* https://github.com/actions/virtual-environments/blob/main/images/linux/Ubuntu2004-Readme.md