Handle PHP 8.1 "never" return type
PHP 8.1 added a "never" return type that indicates that the function
will never return, and raises a fatal error if such a function includes
a `return`. We need to detect and handle such functions in a similar
manner as we handle "void".

Also it turns out that the error handling for a "void" redefinition
trying to return a value was broken, giving an error like
```
Fatal error: Uncaught Error: Undefined constant "thrownew\Patchwork\Exceptions\NonNullToVoid" in file line #
```
so I added a test for that too and adjusted the condense function to
collapse whitespace to a single space instead of stripping it
completely.