Remove support for HHVM
HHVM dropped support for PHP at the end of 2018.

While testing and basic support for HHVM was enabled in Patchwork, it has basically been broken since then and the tests aren't being run against HHVM anyway.

So, let's formally drop support for it and remove all references to it.

This commit actions this in the following manner:
* Mark dedicated functions/classes related to HHVM as (soft) deprecated.
* Remove all conditionals in the codebase related to HHVM.

Follow-up actions:
* I suggest opening an issue for `2.last` to hard deprecate the functions/classes (i.e. trigger a deprecation notice).
* I also suggest opening an issue for 3.0 to remove the deprecated functions/classes.

Ref: https://hhvm.com/blog/2017/09/18/the-future-of-hhvm.html

Fixes 139