getCachedPath(): improve test
The test as added in PR 170 did not actually hit the PHP 8.4 deprecation notice, which - while it is a useful test for other reasons - makes it an ineffective test for safeguarding the fix from 170, which was about preventing the PHP 8.4 deprecation notice.

This commit updates the test to ensure the fix for the PHP 8.4 deprecation notice is now properly safeguarded.

This can be verified by undoing the change to the `src/CodeManipulation.php` file from 170 and running this test on PHP 8.4. The test will fail and show the deprecation notice.
When the change from 170 is re-applied, the test will pass without the deprecation notice.

Includes adding the `===DONE===` markers for consistency with the other tests.