PHP 8.5 | Prevent some "Using null as an array offset" deprecation notices
These are only some "low hanging fruit" fixes. There is a deeper problem in the underlying wiring causing a lot more of these deprecation notices, which I haven't been able to track down yet.

At least the fixes in this PR should allow to focus on the deeper problem by getting rid of a sizable amount of deprecation notices.

Ref: https://wiki.php.net/rfc/deprecations_php_8_5#deprecate_using_values_null_as_an_array_offset_and_when_calling_array_key_exists